provide and details page added

+++ src/actions/cart.action.ts
@@ -1,5 +1,6 @@
 "use server"
 
+import { userService } from "@/app/services/userService";
 import { env } from "@/env";
 import { revalidatePath } from "next/cache";
 import { cookies } from "next/headers";
@@ -7,7 +8,10 @@ const API_URL = env.API_URL
 
 
 export async function deleteCartItem(id: string) {
-    console.log(id);
+    const session = await userService.getSession()
+    if (session.data.user.role !== "CUSTOMER") {
+        return "Only Customer Can Add Product to Cart"
+    }
     const cookieStore = await cookies();
     const res = await fetch(`${API_URL}/cart/${id}`, {
         method: "DELETE",
@@ -21,6 +25,9 @@ export async function deleteCartItem(id: string) {
 }
 // Addd Cart
 export async function addToCart(mealId: string, quantity: number) {
+
+    const session = await userService.getSession()
+
     const cookieStore = await cookies();
 
     const res = await fetch(`${API_URL}/cart`, {
